back/requirements.txt
django == 4.2.5
pytest == 8.2.2
djangorestframework == 3.15.2
django-cors-headers == 4.4.0
requests == 2.32.3
psycopg2-binary == 2.9.9
drf-spectacular == 0.27.2
pillow == 10.4.0